update for RUNTIME==local

+++ clear_env.sh
@@ -1,5 +1,7 @@
 CURRENT_DIR=`dirname $0`
 source ${CURRENT_DIR}/env.sh
 
-adb shell rm -rf $PHONE_DATA_DIR
-rm -rf codegen/models codegen/opencl codegen/tuning
+if [ x"$RUNTIME" != x"local" ]; then
+  adb shell rm -rf $PHONE_DATA_DIR
+  rm -rf codegen/models codegen/opencl codegen/tuning
+fi
